Add files via upload
A dataset from Kaggle regarding campus recruitment. This data set consists of Placement data of students in a specific campus. It includes secondary and higher secondary school percentage and specialization. It also includes degree specialization, type and Work experience and salary offers to the placed students. We can use different algorithms to fiddle with the data and predict Salary, Placement, or even Gender and Family Background.