Add files via upload
This data set consists of Placement data of students in a specific campus. It includes secondary and higher secondary school percentage and specialization. It also includes degree specialization, type and Work experience and salary offers to the placed students. We can use different algorithms to fiddle around with the data and predict Salary, Placement Status, and even Family Background.